requirements.txt
attrs==22.2.0
iniconfig==2.0.0
packaging==23.0
pluggy>=1.0.0,<2.0.0
psycopg2-binary==2.9.9
PyQt5==5.15.10
pandas==2.2.2
pyobjc==10.3.1
